env variable changes

+++ cypress/utils/clean.js
@@ -1,9 +1,28 @@
-const fs= require('fs-extra');
-fs.emptyDir('cypress/reports');
-fs.emptyDir('cypress/screenshots');
-fs.emptyDir('cypress/videos');
-fs.emptyDir('reports/api');
-fs.remove('allure-report')
-fs.remove('allure-results')
+const fs = require("fs-extra");
 
-console.log('Cleaned----->>')
+if (fs.existsSync("reports")) {
+  fs.remove("reports");
+}
+if (fs.existsSync("cypress/reports")) {
+  fs.remove("cypress/reports");
+}
+if (fs.existsSync("screenshots")) {
+  fs.remove("screenshots");
+}
+if (fs.existsSync("cypress/screenshots")) {
+  fs.remove("cypress/screenshots");
+}
+if (fs.existsSync("videos")) {
+  fs.remove("videos");
+}
+if (fs.existsSync("cypress/videos")) {
+  fs.remove("cypress/videos");
+}
+if (fs.existsSync("allure-report")) {
+  fs.remove("allure-report");
+}
+if (fs.existsSync("allure-results")) {
+  fs.remove("allure-results");
+}
+
+console.log("Cleaned----->>");
